# Add benchmark, optimized version and safe API.