No rush.

   
```
    validate                time:   [24.114 ns 24.224 ns 24.352 ns]
                            change: [-93.672% -93.578% -93.502%] (p = 0.00 < 0.05)
                            Performance has improved.
    Found 11 outliers among 100 measurements (11.00%)
      1 (1.00%) low severe
      1 (1.00%) low mild
      8 (8.00%) high mild
      1 (1.00%) high severe
    
    checksum                time:   [15.117 ns 15.181 ns 15.244 ns]
                            change: [-95.164% -95.124% -95.081%] (p = 0.00 < 0.05)
                            Performance has improved.
    Found 6 outliers among 100 measurements (6.00%)
      4 (4.00%) low mild
      1 (1.00%) high mild
      1 (1.00%) high severe
```