@jeffcarp 2021-12-11
Hi there, thanks for your patience. Turning this module into a wrapper of another similar module doesn't seem quite right to me. If you're interested in the `luhn` namespace on crates.io I'm happy to move this one to another name.