@pacak 2021-12-12
This works as well. While making a wrapper I wanted to give current users the same API and a free performance improvement at the same time allowing me to keep iterating on the design - 5 iterations later it's twice as fast.