@jeffcarp 2021-12-19
Actually apologies - it looks like it's [not possible to rename crates on crates.io](https://doc.rust-lang.org/cargo/reference/publishing.html#cargo-yank) (it's not a bug it's a feature 🙂). 

> I wanted to give current users the same API

I appreciate the effort but there are [effectively few to none users of this module](https://crates.io/crates/luhn) - so no need to migrate anyone here. For production applications I'm sure users will want to use a more robust/faster module that is actively maintained - so better for them to pick one of those (e.g. `luhn3`). Unfortunately I don't have the bandwidth to provide maintenance for this module - I'll add a caution to the readme to that effect.