@pacak 2021-12-19
You should be able to add me as an owner (same link as you posted) and I'll push version 2.0.0 that replaces everything including link to the repository then will deal with luhn3 in some way